Add comments to csv to mindmap app
