+++ examples/csv-to-mindmap/src/csv-utils.ts
